Physics Implementation 7 - Leaps and Bounds
Added a physics settings file, this stores all values that are settings
specific for the application
Currently has a bounding box minimum size value, this is application
specific
Currently has a bounding box order value, which defines how many levels
of bounding boxes there are in any one chunk of bounding boxes, this is
application specific

AABB
Added a method to return the min of the AABB
Added a method to return the max of the AABB
Fixed an incorrect value in Diagonal() (used to reference points[2]
which AABBs don't have as they only have 2 points, this min, and max of
the AABB)

PhysicsBoundingChunk
Massively changed
Now supports the settings values, and has all the logic necessary to
build the bounds according to the size and depth defined.
Added several helper members and methods, to aid in easily navigating
the new dynamic bounds
Broke GetNearestBodies() will fix next time.
Using an experimental version of AddBody() (not entirely sure it will
work as intended, also aware it is not as elegant, or smart, as it could
be, planned to be worked on)

TODO: migrate Physics to its own project.